Moved methods from preprocess_setup_launch and fsc to utils for clarity & simplicity. Added instructional comments.
